Update dependencies section of Readme
Gulp isn't even being used on the project.
Bower and Sails are listed in package.json file, so it's installed
locally on the project. There is no need to install it globally.
Since in some cases the bower dependencies are not installed properly,
the script bower-deps were added to use the local installation of bower.