For empty lists, use null instead of empty string
When updating a plugin configuration, if the element is an array
and the update is to remove all elements from that list,
Konga was sending an empty string to represent the empty list.

When using the CORS plugin, the plugin code checks for the list to be
null, but not always for an empty list.

This patch makes Konga send a null instead of an empty string
when updating a plugin list to an empty list.

Signed-off-by: Marc Khouzam <marc.khouzam@ville.montreal.qc.ca>